Fix class name handling in the confusion matrix calculation
A bug existed that was seen when class names appeared after the 'world'
label in a sorted set of class names. This then caused out of bounds
iteration errors when calculating the final confusion matrix.

This has been fixed by creating a sorted list of class names after having
removed the 'world' label. In addition, another list is obtained by
appending the 'world' label.

Test Plan:
To ensure that this bug was tested the class names in the
test_example_4 data have been adjusted so that they would appear after
the default 'world' label when sorted. The tests have also been
strengthened so that a more precise solution could be found.
All test pass with this new change.

PDM-156
Change-Id: I16ba855a775ba0659c2ef0df0bd78d0bd5700475