Improved AUC computation when monotonicity cannot be guaranteed
The auc function checks that both x and y are monotonically
increasing. This is generally the case, but when alpha_st > 0
such property can not be guaranteed hence, with this change, the
check is relaxed for this particular case. Such change is also
documented in the README file.

Minor cleanup: the eTPR was computed twice unnecessarily so the
_psds internal function has been removed without affecting the
psds behaviour.

Test Plan:
 * new test added for alpha_st negative value
 * new test added for AUC behaviour when alpha_st > 0

Change-Id: I85c071f06888d972155ef33bebb3924ff887ad56