Fix confusion matrix so that the axis follow the convention
The confusion matrix (i.e. counts) was built with inverted axis.
Now it follows the convention of having system detections on the
first axis (rows) and ground truths on the second axis (columns).

The docstring for such function has been updated to specify
the convention and the tests were adapted.

Test Plan:
 * Update the expected results for the examples and re-run it

Ref: TI-115

Change-Id: If327c690db6cc426ad0f1192127c306d4ab7707e