Adding empty input file validation
This change makes it impossible to run the code with empty label files.
This has been done by improving the input table validation so that an
error is raised when an empty table is provided. This saves a user
having to debug complex errors that are caused by empty input
tables/files.

PDM-547

Test Plan:
I have added tests to check the new desired behaviours and I have
run all the other tests to make sure there is no regression.

Change-Id: I3d5483a93a8ab21cfd74cd1517b5006fa934a55b