Choose op. points with given criteria (ex. best f-score).
This patchset adds two main features:

- Allows a user to insert information per operating point. A
dictionary can be passed as the "info" parameter when adding an operating
point. This information will be kept in the operating_points table,
and will serve as a reference to how that operating point was created.

- Allows a user to retrieve operating points based on given criteria.
The function "select_operating_points_per_class" can be used to
extract the operating point characteristics, and the corresponding
"info" dictionary if given, that correspond to the requested constraints
on TPR, FPR, eFPR or f-score for an indicated class.

Further changes:
- The example script and the jupyter notebook have been updated with
the select_operating_points_per_class function
- The Readme has been updated to inform the users of the new
functionality.
- The hashing algorithm to compare the input detection tables is updated
so that the hashing is performed after the sorting of the rows and columns.
This allows recognizing identical tables even when the rows and columns are
shuffled.

TI-175

Test Plan:
6 new unit tests have been added to test the new functionalities.
3 of these test the info dictionary addition and behavior, and
the other 3 tests check the retrieval of operating points.

Change-Id: Id97697c15cbf7514801189eb3bb353481e725b73