Ground Truth and Metadata tables drop extra columns
To ensure that .ground_truth and .metadata tables are compatible with
methods such as .compute_macro_f_score() this change makes sure that
PSDSEval drops additional columns that may be present in the provided
metadata and ground_truth arrays.

Solves: #4 (PLAT-1097)

Test Plan:
I used the code presented in issue #4 and added a test that provides
PSDSEval with additional metadata and ground_truth columns which
cause an exception before the change and now pass.
All other existing tests pass.

Change-Id: Id2128a8c76024bc96d61c98c9821bf50e4850d42